+++ src/application/signing_keys/traits.rs
@@ -1,4 +1,5 @@
-use crate::{domain::model::SigningKeys, error::Result};
+use crate::domain::model::{SigningKey, SigningKeys};
+use crate::error::Result;
 
 pub trait SigningKeysServiceTrait {
     fn insert_signing_keys(
@@ -7,5 +8,8 @@ pub trait SigningKeysServiceTrait {
         public: &str,
     ) -> impl Future<Output = Result<String>>;
 
-    fn find_signing_key(&self) -> impl Future<Output = Result<(String, SigningKeys)>>;
+    fn load_active_key(&self) -> impl Future<Output = Result<(String, SigningKey)>>;
+
+    fn find_signing_keys(&self) -> impl Future<Output = Result<Vec<SigningKeys>>>;
+    fn rotate_signing_key(&self) -> impl Future<Output = Result<()>>;
 }